examples/requirements.txt
haystack-ai
ollama-haystack
datasets>=2.6.1
pgvector-haystack
python-dotenv